requirements.txt
telethon
tqdm
python-dotenv
